Fixed small bug
